seperate navbar from homepage
