+++ src/app/front/home/HomePage.tsx
@@ -1,67 +1,14 @@
 'use client';
 
-import React, { useState, useEffect } from 'react';
 import Image from 'next/image';
-import { Button } from '@/components/ui/button';
+import Navbar from './NavBar';
 
 const HomePage = () => {
-	const [isScrolled, setIsScrolled] = useState(false);
-
-	// Function to track scroll position
-	useEffect(() => {
-		const handleScroll = () => {
-			// Change navbar when scrolled past 100px (adjust as needed)
-			if (window.scrollY > 100) {
-				setIsScrolled(true);
-			} else {
-				setIsScrolled(false);
-			}
-		};
-
-		// Listen for scroll events
-		window.addEventListener('scroll', handleScroll);
-		return () => window.removeEventListener('scroll', handleScroll);
-	}, []);
-
 	return (
 		<div className='min-h-screen flex flex-col'>
-			<nav
-				className={`fixed top-0 w-full z-20 py-6 transition-all duration-300 ${
-					isScrolled
-						? 'bg-white/30 backdrop-blur-md shadow-lg'
-						: 'bg-transparent'
-				}`}
-			>
-				<div className='container mx-auto px-6 flex justify-center items-center md:gap-20 sm:gap-10'>
-					<Button
-						variant='outline'
-						className='rounded-full px-6 py-5 text-xl font-bold border-2 text-blue-500 hover:bg-blue-500 hover:text-white active:bg-blue-600 active:text-white transition duration-300'
-					>
-						Home
-					</Button>
-
-					<Button
-						variant='outline'
-						className='rounded-full px-6 py-5 text-lg font-bold border-2 text-blue-500 hover:bg-blue-500 hover:text-white active:bg-blue-600 active:text-white transition duration-300'
-					>
-						Users
-					</Button>
-
-					<Button
-						variant='outline'
-						className='rounded-full px-6 py-5 text-lg font-bold border-2 text-blue-500 hover:bg-blue-500 hover:text-white active:bg-blue-600 active:text-white transition duration-300'
-					>
-						Public
-					</Button>
-
-					<Button
-						variant='outline'
-						className='rounded-full px-6 py-5 text-lg font-bold border-2 text-blue-500 hover:bg-blue-500 hover:text-white active:bg-blue-600 active:text-white transition duration-300'
-					>
-						Contacts
-					</Button>
-				</div>
-			</nav>
+			<header>
+				<Navbar />
+			</header>
 
 			<section className='relative flex flex-col items-center justify-center min-h-screen text-center'>
 				{/* background image  */}
